backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose==3.3.0
hatchet-sdk==0.31.0
python-multipart==0.0.6
aiohttp==3.9.1
openai==1.57.4
google-generativeai==0.8.3
PyPDF2==3.0.1
PyMuPDF==1.23.8
Pillow==10.2.0
